requirements.txt
mcp>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 